Non working test - will work if lobby deals with GameOccassion

+++ api/src/main/java/se/bupp/cs3k/api/GameServerFacade.java
@@ -17,9 +17,9 @@ public interface GameServerFacade {
 
     //public void setScore(Integer gameSessionId, String serializedScore) throws RemoteException;
 
-    public void startGame(Integer gameSessionId, Map<Integer,Integer> teamsByBlayers) throws RemoteException;
+    public void startGame(Long gameSessionId, Map<Integer,Integer> teamsByBlayers) throws RemoteException;
 
-    public void startGame(Integer gameSessionId, List<Integer> players) throws RemoteException;
+    public void startGame(Long gameSessionId, List<Integer> players) throws RemoteException;
 
-    public void endGame(Integer gameSessionId, String serializedScore) throws RemoteException;
+    public void endGame(Long gameSessionId, String serializedScore) throws RemoteException;
 }
